+++ js/scripts.js
@@ -1,6 +1,8 @@
 import im from '../node_modules/include-media-export/include-media.js';
 import fitvids from '../node_modules/fitvids/dist/fitvids.js';
 
+import './_svgs.js';
+
 (function() {
 	'use strict';